backend/pyproject.toml
[project]
name = "intric"
version = "0.1.0"
description = ""
authors = [{ name = "Sundsvalls Kommun" }]
requires-python = ">=3.11,<4"
dependencies = [
    "fastapi>=0,<1",
    "uvicorn[standard]>=0.30.0,<1",
    "openai~=1.4",
    "python-dotenv>=1.1.0",
    "pydantic[email]>=2,<3",
    "tenacity>=8.2.1,<9",
    "tiktoken>=0,<1",
    "numpy>=1.24.2,<2",
    "python-pptx>=1.0.2,<2",
    "arrow>=1.2.3,<2",
    "alembic>=1.10.2,<2",
    "asyncpg>=0.27.0,<0.28",
    "psycopg2-binary>=2.9.5,<3",
    "pyjwt>=2.6.0,<3",
    "black>=24.3.0,<25",
    "fpdf2>=2.8.5,<3",
    "python-docx>=1.2.0,<2",
    "pycryptodome~=3.17",
    "sse-starlette>=1.6.1,<2",
    "gunicorn>=23.0.0,<24",
    "psutil>=5.9.5,<6",
    "pdfplumber>=0.11,<1",
    "langchain>=0.3.0,<0.4",
    "jsonschema>=4.19.2,<5",
    "pydantic-settings>=2.1.0,<3",
    "wrapt>=1.16.0,<2",
    "python-multipart>=0.0.18,<0.0.19",
    "arq>=0.26,<1",  # v0.26+ required for allow_abort_jobs, lifecycle hooks
    "celery[redis]>=5.4,<6",
    "soundfile>=0.12.1,<0.13",
    "audioread>=3.0.1,<4",
    "pydantic-extra-types>=2.6.0,<3",
    "bcrypt>=4.1.2,<5",
    "dependency-injector>=4.41.0,<5",
    "docx2python>=3.5.0,<4",
    "jinja2>=3.1.5,<4",
    "scrapy>=2.11.2,<3",
    "html2text>=2024.2.26,<2025",
    "beautifulsoup4>=4.12.3,<5",
    "crochet>=2.1.1,<3",
    "pgvector>=0.2.5,<0.3",
    "anthropic>=0.28.1,<0.29",
    "sqlalchemy-mixins>=2.0.5,<3",
    "python-magic>=0.4.27,<0.5",
    "aiocache>=0.12.2,<0.13",
    "redis>=5.0.8,<6",
    "sentry-sdk[fastapi]==2.8.0",
    "mistralai>=1.5.0,<2",
    "tavily-python>=0.5.1,<0.6",
    "greenlet>=3.2.2,<4",
    "litellm>=1.76.1,<2",
    "mcp>=1.2.0,<2", # Model Context Protocol SDK for MCP server integration
    "cryptography>=43.0.0,<44", # For Fernet encryption of HTTP auth passwords
    "httpx~=0.28.0",
    "pandas>=2.2.3,<3",
    "orjson>=3.9,<4",
    "aiofiles>=23.0,<24",
    "python-calamine>=0.3.1,<1",  # Fast Excel reader engine for pandas
    "rich>=14.2.0"
]

[project.scripts]
start = "intric.server.main:start"

[dependency-groups]
dev = [
    "pytest>=7.2.1,<8",
    "pytest-cov>=4.0.0,<5",
    "pytest-asyncio>=0.20.3,<0.21",
    "asgi-lifespan>=2.1.0,<3",
    "pyright>=1.1.390,<2",
    "ruff>=0.12.0,<0.13",
    "aioresponses>=0.7.4,<0.8",
    "freezegun>=1.2.2,<2",
    "pytest-integration>=0.2.3,<0.3",
    "pytest-vcr>=1.0.2,<2",
    "pytest-env>=1.1.3,<2",
    "vcrpy>=5,<6",
    "testcontainers>=4.8.1,<5",
    "pytest-dotenv>=0.5.2,<0.6",
    "openpyxl>=3.1.5,<4",
    "pytest-xdist>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intric"]

[tool.ruff]
line-length = 88
exclude = [
    ".venv",
    "__pycache__",
    ".git",
    ".pytest_cache",
    "build",
    "dist",
    "*.egg-info"
]

[tool.ruff.lint]
# Minimal ruleset to start with
select = [
    "F",   # Pyflakes (correctness: undefined names, etc.)
    "E9",  # pycodestyle “runtime”/syntax errors,
]

# Basic ignores for formatter compatibility (kept for when E/W are enabled later)
ignore = []

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]  # Allow re-exports without being flagged as unused

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
